rename dashboard to help, notifications to profile
